move useEventClock to event.services
