+++ client/src/pages/event/event.fetch.js
@@ -1,4 +1,4 @@
-import { useState, useEffect } from "react";
+
 import { 
   fetchApi, getTags,
   useEventQuery, usePlayerQuery,
@@ -7,7 +7,7 @@ import {
 import { useMatchQuery } from '../match/match.fetch';
 import { useSetEventMutation } from '../eventEditor/eventEditor.fetch';
 import { nextRoundUpdate, clearRoundUpdate, clockUpdate } from './services/eventFetch.services'
-import { calcClock, clockPoll, isZero } from './services/clock.services';
+import { calcClock } from './services/clock.services';
 import { debugLogging } from '../../assets/config';
 
 export const eventApi = fetchApi.injectEndpoints({
@@ -53,17 +53,5 @@ export const eventApi = fetchApi.injectEndpoints({
 const refetchStats = (id) => fetchApi.util.invalidateTags(getTags('Stats',{all:0})({id}))
 
 
-export function useEventClock(id, status = 0, eventClock = null, defaultPoll = 0) {
-    const [pollingInterval, setPoll] = useState(defaultPoll);
-    const { data, error, isLoading } = eventApi.useClockQuery(id, { skip: !id, pollingInterval })
-  
-    useEffect(() => {
-      if (data) {
-        setPoll(isZero(eventClock) ? 0 : clockPoll(data.state, status))
-      }
-    }, [data, status, eventClock]);
-    return { data, error, isLoading }
-}
-
 export { useEventQuery, usePlayerQuery, useSettingsQuery, useStatsQuery, useSetEventMutation, useMatchQuery, refetchStats };
-export const { useNextRoundMutation, useClearRoundMutation, useUpdateCreditsMutation, useClockActionMutation } = eventApi;
+export const { useNextRoundMutation, useClearRoundMutation, useUpdateCreditsMutation, useClockActionMutation, useClockQuery } = eventApi;
